Added skeletal data fetching service
The interaction between asynctasks and screen rotation (and other config
changes) makes me strongly prefer a simple service for data fetching.

Currently the class does nothing, but it will eventually fetch data in the
background and then respond to activities via a broadcast.  That way screen
rotation and company are handled gracefully and without repeating actions.