uri scheme now 'data' not 'exchangedata'
Since FetchService uses intents as a sort of asynchronous communication
unpegged to the activity lifecycle, it uses it's own specific intent actions
and thus doesn't need a unique uri scheme.

ACTION_VIEW almost makes sense for the receiving of data, but it would be
bizarre for Android to ask what program to view the data in when a user just
opened this program.