FetchServiceTest now works with multiple tests
Because there's no way to have a test case wait for a service to stop, and
there's no way to ask for a fresh context between test runs, race conditions
abound when running more than one test since the broadcast intent of one test
run could trigger several others.

To combat this, a ServiceTestCase base class is added that has a joinService()
method that behaves much like the Thread.join() in that it waits for the
service to end.  Sadly it's a busy wait since I don't think it's possible to
block on a service like a Thread.